split tests for contract/compiler into different files
